Reviewed and fixed conflicts - confirming merge.

+++ Exercise1/src/alarm.c
@@ -1,8 +1,17 @@
 #include "alarm.h"
 
+#include <time.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <time.h>
+#include <string.h>
+#include <signal.h>
+#include <dirent.h>
+
 //Initial alarm count is zero
 unsigned int alarm_count = 0;
 alarm_t alarms[MAX];
+int status;
 
 alarm_t get_alarm(unsigned int id)
 {
@@ -22,7 +31,6 @@ int get_free_alarm_id()
 {
     for(int i = 0; i<MAX; i++)
     {
-        printf("%i:  %i  \n", i,  alarms[i].active);
         if(alarms[i].active == 0)
         {
             return i;
@@ -33,7 +41,7 @@ int get_free_alarm_id()
 }
 
 //Helper Method: After entering the date and time for an alarm, create a new child process using fork(2)
-unsigned int _fork(time_t timestamp)
+unsigned int _fork(time_t timestamp, char * trackpath, int tracknum)
 {
     unsigned int pid = fork();
 
@@ -46,7 +54,7 @@ unsigned int _fork(time_t timestamp)
         sleep(timestamp - time(NULL));
 
         //Playing alarm sound
-        play_alarm_sound();
+        play_alarm_sound(trackpath, tracknum);
 
         // clean up, setting the alarm to inactive, and exiting the process
         exit(pid);
@@ -54,7 +62,7 @@ unsigned int _fork(time_t timestamp)
     }
 }
 
-unsigned int schedule_alarm(time_t timestamp)
+unsigned int schedule_alarm(time_t timestamp, char * trackpath, int tracknum)
 {
     int alarm_id = get_free_alarm_id();
     if(alarm_id != -1)
@@ -65,7 +73,7 @@ unsigned int schedule_alarm(time_t timestamp)
         new_alarm.active = 1;
 
         //Child
-        unsigned int pid = _fork(timestamp);
+        unsigned int pid = _fork(timestamp, trackpath, tracknum);
         new_alarm.pid = pid;
 
         //Designates an ID equal to the alarm count
@@ -83,13 +91,30 @@ unsigned int schedule_alarm(time_t timestamp)
 void cancel_alarm(unsigned int id)
 {
     // The alarm needs to be removed from the list after being canceled, at this time nothing happens to it.
-    alarm_t alarm = get_alarm(id);
-    kill(alarm.pid, SIGKILL);
-    // Set the alarm as inactives
+    if(0 <= id && id < MAX)
+    {
+        alarm_t alarm = get_alarm(id);
+        if(alarm.pid > 0)
+        {
+            kill(alarm.pid, SIGKILL);
+            // Set the alarm as inactives
+            alarms[id].active    = 0;
+            alarms[id].pid       = 0;
+            alarms[id].timestamp = 0;
+            // updating alarm count
+            alarm_count -= 1;
+            printf("test id: %c", id);
+        }
+        else
+        {
+            printf("The alarm at the given index %ui does not have a valid PID, no alarms canceled.\n", id);
+        }
+    }
+    else
+    {
+        printf("Index %ui is out of bounds, no alarms canceled.\n", id);
 
-    alarms[id].active = 0;
-    alarm_count -= 1;
-    printf("test id: %c", id);
+    }
 }
 
 void list_active_alarms()
@@ -101,18 +126,93 @@ void list_active_alarms()
         {
             printf("ID: %i - PID: %u - ACTIVE - TIME: %s", i, alarms[i].pid, ctime(&alarms[i].timestamp));
         }
-        else
-        {
-            printf("ID: %i - PID: %u - NOT ACTIVE - TIME: %s", i, alarms[i].pid, ctime(&alarms[i].timestamp));
-        }
     }
     printf("END OF LIST. \n");
 }
 
 
-void play_alarm_sound()
+void list_available_alarm_tracks(char * path)
 {
-    if(SYSTYPE == 2) system("mpg123 --quiet ringtones/sea_shanty.mp3");
-    else if (SYSTYPE == 1) system("afplay ringtones/sea_shanty.mp3");
+    DIR *dir;
+    struct dirent *ent;
+    if ((dir = opendir (path)) != NULL) {
+        /* print all the files and directories within directory */
+        int counter = 0;
+        printf("TRACK NUM   TRACK NAME \n");
+        while ((ent = readdir (dir)) != NULL) {
+            if(strcmp("..", ent->d_name) != 0 && strcmp(".", ent->d_name) != 0 )
+            {
+                printf ("    %i       %s \n", counter, ent->d_name );
+                counter ++;
+            }
+        }
+        closedir (dir);
+    } else {
+        /* could not open directory */
+        perror ("");
+    }
+}
+
+
+int get_available_alarm_tracks(char * path, char ** tracklist)
+{
+    DIR *dir;
+    struct dirent *ent;
+
+    if ((dir = opendir (path)) != NULL) {
+        /* print all the files and directories within directory */
+        int counter = 0;
+        while ((ent = readdir (dir)) != NULL && counter < MAX_TRACKS) {
+            if(strcmp("..", ent->d_name) != 0 && strcmp(".", ent->d_name) != 0 )
+            {
+                // This is probably horrible and leaky...
+                tracklist[counter] = malloc((strlen(path) + strlen(ent->d_name) + strlen("/") + 1)*sizeof(char));
+                strcpy(tracklist[counter], path);
+                strcat(tracklist[counter],"/");
+                strcat(tracklist[counter], ent->d_name);
+                counter ++;
+            }
+        }
+        closedir (dir);
+        return counter;
+    } else {
+        /* could not open directory */
+        perror ("");
+        return -1;
+    }
+}
+
+void play_alarm_sound(char *path, int index)
+{
+    int trackcount;
+    char ** tracklist = calloc(MAX_TRACKS, sizeof(char*));
+    // this should rather be stored as a list in the main function, than being run over again.
+    trackcount = get_available_alarm_tracks(path, tracklist);
+    if(index >= 0 && index < trackcount)
+    {
+        if(SYSTYPE == 2)
+        {
+            char command[1000] =  "mpg123 --quiet ";
+            strcat(command, tracklist[index]);
+            system(command);
+        }
+        else if (SYSTYPE == 1)
+        {
+            char command[1000] ="afplay --quiet ";
+            strcat(command, tracklist[index]);
+            system(command);
+        }
+        else
+        {
+            printf("RING RING! This program does not support audio alarms on your system. \n");
+        }
+    }
+    else
+    {
+        printf("RING RING! Provided index is out of range.\n");
+    }
+    // freeing the allocated memory
+    for(int i=0; i<MAX_TRACKS;i++) free(tracklist[i]);
+    free(tracklist);
 }
 //2023-01-10 13:37:00
